feat: Diamond manager class
Diamond Manager class keeps track of current diamonds in the scene.
This will be used to manipulate diamonds in the scene. Power ups are
one example of how to manipulate current diamonds in the scene. It gets
nearby diamonds by calculating distance from a (x,y) and the diamond (x,y).