Add new Power Up feature
Hero can now engage with powerups. When power up is active, the hero attracts diamonds within a configured radius from itself.
Animation added to the diamonds so they attract and move closer to the heros position with gsap.to(). When using gsap.to() I had to keep track of each tween applied to each diamond thats attracted. In doing so Hero class must keep track of each diamond and its tween in order to destroy both. Gsap causes errors if tween is not killed.